Fix toggling of widget settings
A major change was made to the widgets API such that the reducer for the
WidgetContents component must only modify the widget's data object. It
cannot modify the top-level widget object. This is to prevent the state
from overlapping with the reducer in the Widget component itself.

In other words, there can be only one reducer that modifies the
top-level widget object.