Correct placeholder color once and for all
The original placeholder color tweak in 8684ac9 was a necessary one,
because inputs with white backgrounds needed to have dark placeholder
colors. However, this tweak caused a regression in the placeholder color
for note widgets. Therefore, we fix this issue once and for all by
setting two separate placeholder colors for the respective selectors.