Fix fetcher logic by resetting fetchError when setting query
There are two constraints that require this:
1. fetchError must be specified as a condition on invoking another fetch
call, lest we trigger an infinite fetch loop
2. If we do the above, we cannot submit a valid query after an invalid
query was attempted *unless* we reset fetchError when the query is set.

(2) isn't my favorite approach, since it requires the sub-reducer (i.e.
the reducer of the implemented widget) to know when to set fetchError,
which it shouldn't have to.

To work around this, I need to reset the fetch error immediately before
resetting the query (in the form submit handler where the query is being
set). Performance should not be a concern, since this code will all take
place within a React event handler, so the two dispatch calls should
only trigger a single (batched) render.