Strictly limit when the request query can be fetched
This fixes a longstanding design flaw where not having all of the proper
conditions in place could cause an infinite fetch loop. Now, that is no
longer possible, because the useEffect-based fetch logic has been
constrained to run only once. Otherwise, it must run on user action
(therefore making it impossible to run more than once in succession).
This also eliminates the need for the dependencies array, as well as the
need for the conditions within the useEffect call (since now, the
useEffect call only runs once when the widget initially loads).