Fix audio seeker not updating as user dragged seeker
This bug was due to a regression in 6e16798 (and by extension, 6040f68)
that caused the timeupdate listener to latch onto the original
value of currentTime, and therefore wasn't responding to updates because
the aforementioned commits limited the listener to only being bound
once.

This new commit means reverting to a behavior where the events are bound
and unbound every second of audio, but for now, this is what needs to
happen in order for the state's latest currentTime value to be read.