Prefer useMemo for debounce instead of useCallback
Per facebook/react#19240, useMemo is the recommended hook for consuming
a debounced or throttled function. This is because useMemo gives the
component access to the original (cancelable) return values of
debounce() and throttle(), with their added cancel() methods. In
contrast, we lose access to the cancel() methods via useCallback. For
more information, see:
https://github.com/facebook/react/issues/19240#issuecomment-652945246
https://github.com/facebook/react/issues/19240#issuecomment-867885511