Remove logic to listen for widget changes on the server
This fixes an issue with updating the widget state after it was just
pushed to the server on the same device, and overall just makes things
simpler. The caveat is that if you are on one device, you will need to
reload the page to see changes you just made on another device.