Do not fetch Podcast results on initial page load
This improves the performance of the app such that the Podcast widget
does not push its state to the server on initial page load (even if
nothing has changed). It also solves a major hurdle with testing the
sync functionality; the refactored tests will be coming in the next
commit.