Make sync tests atomic
Previously, one failing sync test caused other sync tests to fail. This
was because a failing assertion in one test would prevent the
mockRestore() calls from running at the end of the failing test. Thus,
the mocks were not properly reset for subsequent tests.

To fix this, we moved the mock teardown logic to the afterEach() scope,
so that the mocks will be properly reset regardless of whether a test
passes or fails. This is similar to how a try..finally is used to ensure
that teardown logic is always run, regardless of any exceptions thrown
beforehand.