Fix CORS errors when sending Supabase API requests
This was due to the massive CAPTCHA response strings being written to
the user metadata store, which gets serialized into the JWT that's sent
on every request. Hence, the request is too large for the server to
handle.

If we prevent the CAPTCHA responses from ever being written to the user
metadata, we should be good to go.