Fix Google Analytics not loading on front end
This bug was ultimately caused by the refactor in 3bb979d, because
_document.tsx is only rendered on the server-side (meaning a useEffect
hook inside _document will never actually run). But per my testing on my
personal website, it seems we may not need a nonce anyway. Trying
without it to see where we get.

TODO: if we can do without the nonce, make sure to switch back to
getStaticProps to enable static generation / disable server-side
rendering.