Fix JSON serialization in Postgres database
All of the raw_data values (for both the dashboards and widgets tables)
were strings of JSON, rather than proper/queryable JSON objects. This
commit adds a migration script to unwrap the JSON objects, and also
updates the code to eliminate the stringification and string-parsing
that was ultimately causing the problem in the first place.