Add support for email confirmation after sign-up
This functionality was really implemented out of necessity because there
is a bug in Supabase CLI where the latest supabase/gotrue image was not
being pulled, thus leaving the Sign Up submission broken due to the PKCE
error.

However, as the error states, one can enable autoconfirm (i.e.
email confirmations) to resolve the error. This was never ideal, since
we already have the Captcha, and I would prefer to have a frictionless
Sign Up experience. But since the PKCE error is absent on Staging (while
autoconfirm is still disabled there), that means we only need to put up
with autoconfirm on Local if we want to work around the issue for the
time being.

Therefore, this commit implements Email Confirmation functionality. We
still need to write automated tests for it, but testing it manually did
go very well, and the UI/UX is quite satisfactory.