Add missing code to redirect to home after sign-up
Only if email confirmations are disabled. We used to have this logic,
but it got accidentally removed when the email confirmation flow was
implemented back in cd354a5. This new commit adds the redirect logic
back in, and also adds the relevant test assertion to verify the
behavior. This commit also consolidates the logic for mocking the
Location object for much easier reuse across test suites.