Move initial database schema setup to migration scripts
Including all table and function creation. This commit also fixes some
of the SQL syntax and makes the table/policy creation scripts (for
dashboards and widgets) fully idempotent. This will be important for the
eventual deploy to Production because the dashboards and widgets tables
already exists in Production.

Also, it's important to note that the 'dashboards'/'widgets' table
migrations have been manually inserted before all other migrations (this
is why the migration version numbers are so exact). To ensure that the
Staging and Production environments don't have a problem with this,
we've added the --include-all flag when pushing the database to both
Supabase environments (see the relevant GitHub actions in the .github/
directory).